refactor(commands): introduce command-level Config::try_from and tighten field visibility
Lift the argument-level `DatabaseConfig::try_from(&self.db_arguments)`
pattern out of `execute()` and into `impl TryFrom<&StdioCommand> for Config`
and `impl TryFrom<&HttpCommand> for Config`. Each subcommand now produces
a fully-populated `Config` (with `http: None` for stdio, `http: Some(..)`
for http) that gets passed wholesale to `common::create_server`, which
in turn takes `&Config` again and reads `config.database.backend` /
`config.database.read_only`.

Drop `pub(crate)` from `StdioCommand.db_arguments` and
`HttpCommand.db_arguments` — the field is now genuinely private to
its own module. `cli.rs` no longer reaches into either command's
internals, and the `stdio_db` / `http_db` test helpers are gone.

Relocate the tests that exercise `DatabaseArguments` field wiring and
the `TryFrom<&DatabaseArguments> for DatabaseConfig` conversion from
`cli.rs` into `src/commands/stdio.rs`, where they share the module with
the now-private field and can parse a `StdioCommand` directly via
`StdioCommand::try_parse_from(["_", ...])` without going through the
top-level `Arguments` wrapper. Drop the `cli.rs` test module entirely —
every test that remained there was exercising basic clap behavior
(subcommand-required, unknown-argument rejection, enum subcommand
parsing) rather than our custom logic.

Also drop `db_connection_timeout_zero_rejected` and the stand-alone
http test file module for the same reason: they only verified clap's
own `value_parser!(u64).range(1..)` validator and clap's `flatten`
mechanics, not anything we author.
