+++ src/cli.rs
@@ -111,109 +111,3 @@ pub(crate) async fn run() -> Result<(), Error> {
         Command::Http(cmd) => cmd.execute().await,
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::commands::common::DatabaseArguments;
-    use database_mcp_config::{ConfigError, DatabaseBackend, DatabaseConfig};
-
-    #[track_caller]
-    fn parse(args: &[&str]) -> Arguments {
-        Arguments::try_parse_from(args).expect("valid CLI arguments")
-    }
-
-    #[track_caller]
-    fn stdio_db(args: &Arguments) -> &DatabaseArguments {
-        match &args.command {
-            Command::Stdio(cmd) => &cmd.db_arguments,
-            _ => panic!("expected stdio subcommand"),
-        }
-    }
-
-    #[track_caller]
-    fn http_db(args: &Arguments) -> &DatabaseArguments {
-        match &args.command {
-            Command::Http(cmd) => &cmd.db_arguments,
-            _ => panic!("expected http subcommand"),
-        }
-    }
-
-    #[test]
-    fn no_subcommand_is_rejected() {
-        let err = Arguments::try_parse_from([BIN]).expect_err("subcommand is required");
-        assert_eq!(
-            err.kind(),
-            clap::error::ErrorKind::DisplayHelpOnMissingArgumentOrSubcommand
-        );
-    }
-
-    #[test]
-    fn db_backend_before_subcommand_rejected() {
-        assert!(Arguments::try_parse_from([BIN, "--db-backend", "mysql", "http"]).is_err());
-        assert!(Arguments::try_parse_from([BIN, "--db-backend", "mysql", "stdio"]).is_err());
-    }
-
-    #[test]
-    fn db_read_only_defaults_to_true() {
-        let args = parse(&[BIN, "stdio"]);
-        assert!(stdio_db(&args).read_only);
-    }
-
-    #[test]
-    fn db_connection_timeout_zero_rejected() {
-        assert!(Arguments::try_parse_from([BIN, "stdio", "--db-connection-timeout", "0"]).is_err());
-    }
-
-    #[test]
-    fn db_query_timeout_zero_passes_through() {
-        let args = parse(&[BIN, "stdio", "--db-query-timeout", "0"]);
-        let config = DatabaseConfig::try_from(stdio_db(&args)).expect("valid db args");
-        assert_eq!(config.query_timeout, Some(0));
-    }
-
-    #[test]
-    fn db_backend_after_http_subcommand() {
-        let args = parse(&[BIN, "http", "--db-backend", "mysql"]);
-        assert_eq!(http_db(&args).backend, DatabaseBackend::Mysql);
-        assert!(matches!(args.command, Command::Http(_)));
-    }
-
-    #[test]
-    fn stdio_subcommand_db_args() {
-        let args = parse(&[
-            BIN,
-            "stdio",
-            "--db-backend",
-            "postgres",
-            "--db-user",
-            "pg",
-            "--db-name",
-            "app",
-        ]);
-        let db = stdio_db(&args);
-        assert_eq!(db.backend, DatabaseBackend::Postgres);
-        assert_eq!(db.user.as_deref(), Some("pg"));
-        assert_eq!(db.name.as_deref(), Some("app"));
-
-        let config = DatabaseConfig::try_from(db).expect("valid postgres args");
-        assert_eq!(config.backend, DatabaseBackend::Postgres);
-        assert_eq!(config.user, "pg");
-        assert_eq!(config.name.as_deref(), Some("app"));
-    }
-
-    #[test]
-    fn version_subcommand() {
-        let args = parse(&[BIN, "version"]);
-        assert!(matches!(args.command, Command::Version));
-    }
-
-    #[test]
-    fn try_from_database_arguments_propagates_validation_errors() {
-        // SQLite without --db-name must fail validation inside the TryFrom impl,
-        // surfacing `ConfigError::MissingSqliteDbName` to the caller.
-        let args = parse(&[BIN, "stdio", "--db-backend", "sqlite"]);
-        let errors = DatabaseConfig::try_from(stdio_db(&args)).expect_err("sqlite without --db-name must be rejected");
-        assert!(errors.iter().any(|e| matches!(e, ConfigError::MissingSqliteDbName)));
-    }
-}
